Load testing and profiling
Added a profiler to test requests, determined bottle neck to be in the API calls
that validate the OIDC tokens with KeyCloak.

Added a caching service that caches the JWT tokens after they have been
validated by Key cloak to reduce the number of API calls (and thus blocking
requests) that are performed.

Migrated to using percona, which means that we can rip out all of the
cockroachdb code. Added a wrapper around HTTP calls which allow us to rollback
a transaction and then retry them in an effort to deal with WSREP locking
issues with Percona.