This commit address the required models, schemas and resources files needed in order to develop the back-end for the new Bookings feature inside theq application.
In order to accomplish this models and schemas were developed first and tested in terminal to ensure they would
compile properly. Next, test data was inserted into a local MySQL instance to ensure that the models worked.
After this was completed, GET end-points for exams, bookings, invigilators and rooms were created to ensure that
this data could be retrieved. Once this was confirmed, the POST, PUT and DELETE end-points were developed and tested.
Finally, all new files that were created were linted using PyLint to ensure that general python coding standards were
met.