Various fixes for IE11 Comptability
Fixed the edit-group-booking-modal re: issue not allowing submission of invigilator change.  Fixed the hang on the add pesticide exam form where the next button would not switch to enabled on the last step.  fixed issue causing the invigilator dropdown select list (in various places) to be blank when the invigilator is unassigned instead of saying 'unassigned'. several other very minor changes to better support IE11 users.