Fixed Bugs Preventing Office Agenda from Displaying Events
Fixed the issues related to certain exams not showing on the Office Agenda, added support for the current exam/booking data models (sbc staff invigilators), and fixed the issue resulting in non-exam events not displaying.  added a user-friendly message to display when there are no exam events in scope and added the ability to click the offsite icon to display the offsite details in the row details.  Also fixed...
-Bug preventing Contract Invigilator being assigned from BookingModal initially
-Bug Preventing Contract Invigilator being assigned for Challenger Exams
-Bug related to pressing enter key on AddExamModal and Calendar view causing refresh
-Bug related to exam filtering on ExamInventoryTable
-Bug related to Yellow Triangle on ExamInventoryTable