Re-built the Booking Modal as plain CSS div container
Due to issues with IE 11 and the document.activeElement call, Bootstrap Modals were not working in this situation so re-wrote the component as a div box styled to match bootstrap with the necessary logic to replace the modal hooks.  also fixed issues with selection of invigilators and added logic to auto focus on the text input when opening.