Fixed Errors in Branch Agenda
Previously this component was more complex and it relied on an invisible full-calendar instance to consume and parse the event data however this approach was causing issues.  Since the Agenda has recently been simplified and no longer supports multiple views etc., there is no need for full-calendar.  Have re-implemented using only moment.js and eliminated the errors related to full-calendar